LPS-75124 Move CharPipe and UnsyncTeeWriter to io package.
