+++ modules/util/portal-tools-sample-sql-builder/src/test/java/com/liferay/portal/tools/sample/sql/builder/io/CharPipeTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.tools.sample.sql.builder;
+package com.liferay.portal.tools.sample.sql.builder.io;
 
 import com.liferay.portal.kernel.test.SyncThrowableThread;
